+++ crates/hyperqueue/src/common/env.rs
@@ -19,4 +19,5 @@ pub const HQ_INSTANCE_ID: &str = create_hq_env!("INSTANCE_ID");
 pub const HQ_SUBMIT_DIR: &str = create_hq_env!("SUBMIT_DIR");
 pub const HQ_ENTRY: &str = create_hq_env!("ENTRY");
 pub const HQ_PIN: &str = create_hq_env!("PIN");
+pub const HQ_TASK_DIR: &str = create_hq_env!("TASK_DIR");
 pub const HQ_CPUS: &str = create_hq_env!("CPUS");